backend/requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.12.7
Django==2.2.19
python-dotenv==0.21.1
python-telegram-bot==13.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.7
tzlocal==4.2
djangorestframework==3.12.4
django-filter==21.1
djoser==2.1.0
